requirements.txt
fastapi
uvicorn[standard]
python-socketio
drake
